fix exit by press C+d
