+++ srcs/main.c
@@ -13,6 +13,29 @@ void	handler(int signo)
 	rl_redisplay();
 }
 
+void	minishell_loop(char **envp)
+{
+	t_envlist		*elst;
+	t_execdata		*data;
+	char			*line;
+
+	elst = create_envlist(envp);
+	while (1)
+	{
+		line = readline("minishell$ ");
+		if (!line)
+			break ;
+		if (line[0] != '\0')
+		{
+			data = parse_cmd(line, elst);
+			execute_start(data);
+			clear_execdata(data);
+			add_history(line);
+		}
+		free(line);
+	}
+}
+
 int	main(int argc, char **argv, char **envp)
 {
 	if (signal(SIGINT, handler) == SIG_ERR
@@ -24,4 +47,5 @@ int	main(int argc, char **argv, char **envp)
 	minishell_loop(envp);
 	(void)argc;
 	(void)argv;
+	return (g_status);
 }
